% Generated by roxygen2 (4.0.1): do not edit by hand
\name{colRun}
\alias{colRun}
\title{Run Colony}
\usage{
colRun(ids, genotypes, marker_names, random_seed = 1234,
  update_allele_frequency = FALSE, inbreeding = FALSE, length_of_runs = 1,
  full_likelihood_precision = 3, monitor_interval = 10000,
  save_file = FALSE)
}
\arguments{
\item{ids}{vector giving the offspring ids}

\item{genotypes}{data frame with the genotypes. Each row is an individual,
in the same order as for ids. Each pair of columns represents one locus.}

\item{marker_names}{a vector of strings with the marker names}

\item{random_seed}{used by Colony}

\item{update_allele_frequency}{boolean, if \code{TRUE} Colony will update
allele frequencies by taking into account of reconstructed pedigrees.}

\item{inbreeding}{boolean, absence of presence of inbreeding in Colony
analysis}

\item{length_of_runs}{Give a value of 1, 2, 3, 4 to indicate short, medium,
long, very long run.}

\item{full_likelihood_precision}{1/2/3=low/medium/high Precision for
Fulllikelihood}

\item{monitor_interval}{interval between records (number of iterations)}

\item{save_file}{boolean, save the results to a file?}
}
\description{
Run a Colony pedigree reconstruction analysis. Some description
  of the parameters are taken from the ColonyUserGuide.pdf and colony2.dat
  provided with Colony, as downloaded on 2014-09-27.
}
\details{
From the pdf manual accompanying Colony: "Update allele frequency:
  Allele frequencies are required in calculating the likelihood of a
  configuration. These frequencies can be provided by the user (see below)
  or are calculated by Colony using the genotypes in OFS, CMS (optional) and
  CFS (optional). In the latter case, you can ask Colony to update allele
  frequency estimates by taking into account of the inferred sibship and
  parentage relationships during the process of searching for the maximum
  likelihood configuration. However, updating allele frequencies could
  increase computational time substantially, and may not improve
  relationship inference much if the genetic structure of your sample is not
  strong (i.e. family sizes small and evenly distributed, most candidates
  are not assigned parentage). I suggest not updating allele frequencies
  except when family sizes (unknown) are suspected to be large (relative to
  sample size) and highly variable."
}

